requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
dnspython>=2.5.0
tqdm>=4.66.0
beautifulsoup4>=4.12.0
colorama>=0.4.6
asyncio>=3.4.3
werkzeug>=3.0.0
jinja2>=3.1.2
click>=8.1.7
itsdangerous>=2.1.2
blinker>=1.7.0 